Merging dev-v0.10(stable) into main branch for backup

+++ sql_init/balance_transactions.sql
@@ -4,8 +4,8 @@ CREATE TABLE balance_transactions (
     user_id UNIQUEIDENTIFIER NOT NULL,
     endpoint_id UNIQUEIDENTIFIER NOT NULL,
     transaction_date DATETIME2 NOT NULL DEFAULT DATEADD(HOUR, 2, GETUTCDATE()),
-    deducted_amount INT NOT NULL,
-    balance_after INT NOT NULL,
+    deducted_amount DECIMAL(10, 2) NOT NULL,
+    balance_after DECIMAL(10, 2) NOT NULL,
     CONSTRAINT FK_balance_transactions_users FOREIGN KEY (user_id) REFERENCES users(id),
     CONSTRAINT FK_balance_transactions_endpoints FOREIGN KEY (endpoint_id) REFERENCES endpoints(id)
 );
